[fix] 미션 엔티티 변경으로 인한 버그 수정

+++ src/main/java/com/modernfarmer/farmusspring/domain/farmclub/vo/GetMissionPostListVo.java
@@ -3,7 +3,7 @@ package com.modernfarmer.farmusspring.domain.farmclub.vo;
 import com.querydsl.core.annotations.QueryProjection;
 
 public record GetMissionPostListVo(
-        Long stepNum,
+        int stepNum,
         String image
 ) {
     @QueryProjection
